[Fix] CuBIC, SPADE with scipy 1.14.0, deprecated `.A` attribute in `scipy.sparse` matrices` (#636)
* fix deprecated .A attribute on coo_matrix
* make precision for cubic explicit
* ensure float64 precision for data when using scipy.stats.kstat

+++ elephant/cubic.py
@@ -237,5 +237,6 @@ def _kstat(data):
     """
     if len(data) == 0:
         raise ValueError('The input data must be a non-empty array')
-    moments = [scipy.stats.kstat(data, n=n) for n in [1, 2, 3]]
+    # Due to issues with precision, ensure float64 (default) is the precision of the data array. (scipy == 1.14.0)
+    moments = [scipy.stats.kstat(data.astype(np.float64), n=n) for n in [1, 2, 3]]
     return moments
